backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["xuhaojun <xuhaojuntw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.47.0"
tiktoken = "^0.7.0"
fastapi = {extras = ["standard"], version = "^0.115.0"}
slowapi = "^0.1.9"
argparse = "^1.4.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
